refactor: move hooks to standalone files
